refactor: ♻️ encapsulate mutation to create or claim assets
