+++ lib/QueryAndMutation.ts
@@ -1,0 +1,157 @@
+import { gql } from "@apollo/client";
+
+export const QUERY_VARIABLES = (unitsAndCurrencies = false) => {
+  const _query = unitsAndCurrencies
+    ? `query {
+    instanceVariables {
+      specs {
+        specCurrency {
+          id
+          name
+        }
+        specProjectDesign {
+          id
+          name
+        }
+        specProjectProduct {
+          id
+          name
+        }
+        specProjectService {
+          id
+          name
+        }
+      }
+      units {
+        unitOne {
+          id
+        }
+      }
+    }
+  }`
+    : `query {
+    instanceVariables {
+      specs {
+        specProjectDesign {
+          id
+          name
+        }
+        specProjectProduct {
+          id
+          name
+        }
+        specProjectService {
+          id
+          name
+        }
+      }
+    }
+  }`;
+  return gql`
+    ${_query}
+  `;
+};
+
+export const CREATE_PROPOSAL = gql`
+  mutation {
+    createProposal(proposal: { name: "price tag", unitBased: true }) {
+      proposal {
+        id
+      }
+    }
+  }
+`;
+export const CREATE_INTENT = gql`
+  mutation ($agent: ID!, $resource: ID!, $oneUnit: ID!, $currency: ID!, $howMuch: Float!) {
+    item: createIntent(
+      intent: {
+        name: "project"
+        action: "transfer"
+        provider: $agent
+        resourceInventoriedAs: $resource
+        resourceQuantity: { hasNumericalValue: 1, hasUnit: $oneUnit }
+      }
+    ) {
+      intent {
+        id
+      }
+    }
+
+    payment: createIntent(
+      intent: {
+        name: "payment"
+        action: "transfer"
+        receiver: $agent
+        resourceConformsTo: $currency
+        resourceQuantity: { hasNumericalValue: $howMuch, hasUnit: $oneUnit }
+      }
+    ) {
+      intent {
+        id
+      }
+    }
+  }
+`;
+
+export const LINK_PROPOSAL_AND_INTENT = gql`
+  mutation ($proposal: ID!, $item: ID!, $payment: ID!) {
+    linkItem: proposeIntent(publishedIn: $proposal, publishes: $item, reciprocal: false) {
+      proposedIntent {
+        id
+      }
+    }
+    linkPayment: proposeIntent(publishedIn: $proposal, publishes: $payment, reciprocal: true) {
+      proposedIntent {
+        id
+      }
+    }
+  }
+`;
+
+export const CREATE_LOCATION = gql`
+  mutation ($name: String!, $addr: String!, $lat: Float!, $lng: Float!) {
+    createSpatialThing(spatialThing: { name: $name, mappableAddress: $addr, lat: $lat, long: $lng }) {
+      spatialThing {
+        id
+        lat
+        long
+      }
+    }
+  }
+`;
+
+export const CREATE_ASSET = (action: "raise" | "transfer" = "raise") => gql`
+  mutation (
+    $name: String!
+    $note: String!
+    $metadata: JSON
+    $agent: ID!
+    $creationTime: DateTime!
+    $location: ID!
+    $tags: [URI!]
+    $resourceSpec: ID!
+    $oneUnit: ID!
+    $images: [IFile!]
+  ) {
+    createEconomicEvent(
+      event: {
+        action: "${action}"
+        provider: $agent
+        receiver: $agent
+        hasPointInTime: $creationTime
+        resourceClassifiedAs: $tags
+        resourceConformsTo: $resourceSpec
+        resourceQuantity: { hasNumericalValue: 1, hasUnit: $oneUnit }
+        toLocation: $location
+      }
+      newInventoriedResource: { name: $name, note: $note, images: $images, metadata: $metadata }
+    ) {
+      economicEvent {
+        id
+        resourceInventoriedAs {
+          id
+        }
+      }
+    }
+  }
+`;
