+++ public/locales/en/ResourceProps.json
@@ -1,5 +1,6 @@
 {
   "assigned to:": "Assigned to:",
   "current location:": "Current location:",
-  "description:": "Here you will see the material passport of asset"
+  "description:": "Here you will see the material passport of asset",
+  "Claim Ownership": "Claim Ownership"
 }
